# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Contract Review App API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/companies/search": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Api Companies Search",
        "operationId": "api_companies_search_api_companies_search_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_CompanySearchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Api Companies Search Get",
        "operationId": "api_companies_search_get_api_companies_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          }
        }
      }
    },
    "/api/companies/{number}": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "Api Company Profile",
        "operationId": "api_company_profile_api_companies__number__get",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "headers": {
              "x-schema-version": {
                "$ref": "#/components/headers/XSchemaVersion"
              },
              "x-latency-ms": {
                "$ref": "#/components/headers/XLatencyMs"
              },
              "x-cid": {
                "$ref": "#/components/headers/XCid"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "default": "/errors/general",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Code"
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Extra"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "_CompanySearchIn": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "_CompanySearchIn"
      },
...
    },
    "headers": {
      "XSchemaVersion": {
        "schema": {
          "type": "string"
        },
        "example": "1.4"
      },
      "XLatencyMs": {
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "example": 12
      },
      "XCid": {
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-f]{64}$"
        },
        "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
      }
    }
  }
}